Redesign class hierarchy
