+++ pyathena/async_cursor.py
@@ -6,8 +6,10 @@ import os
 
 from concurrent.futures.thread import ThreadPoolExecutor
 
+from pyathena.common import CursorIterator
 from pyathena.cursor import BaseCursor
-from pyathena.model import AthenaResultSet
+from pyathena.error import ProgrammingError, NotSupportedError
+from pyathena.result_set import AthenaResultSet
 
 
 _logger = logging.getLogger(__name__)
@@ -19,30 +21,32 @@ class AsyncCursor(BaseCursor):
                  encryption_option, kms_key, converter, formatter,
                  retry_exceptions, retry_attempt, retry_multiplier,
                  retry_max_delay, retry_exponential_base,
-                 max_workers=(os.cpu_count() or 1) * 5):
+                 max_workers=(os.cpu_count() or 1) * 5,
+                 arraysize=CursorIterator.DEFAULT_FETCH_SIZE):
         super(AsyncCursor, self).__init__(client, s3_staging_dir, schema_name, poll_interval,
                                           encryption_option, kms_key, converter, formatter,
                                           retry_exceptions, retry_attempt, retry_multiplier,
                                           retry_max_delay, retry_exponential_base)
         self._executor = ThreadPoolExecutor(max_workers=max_workers)
+        self._arraysize = arraysize
+
+    @property
+    def arraysize(self):
+        return self._arraysize
+
+    @arraysize.setter
+    def arraysize(self, value):
+        if value <= 0 or value > CursorIterator.DEFAULT_FETCH_SIZE:
+            raise ProgrammingError('MaxResults is more than maximum allowed length {0}.'.format(
+                CursorIterator.DEFAULT_FETCH_SIZE))
+        self._arraysize = value
 
     def close(self, wait=False):
         self._executor.shutdown(wait=wait)
 
     def _description(self, query_id):
         result_set = self._collect_result_set(query_id)
-        return [
-            (
-                m.get('Name', None),
-                m.get('Type', None),
-                None,
-                None,
-                m.get('Precision', None),
-                m.get('Scale', None),
-                m.get('Nullable', None)
-            )
-            for m in result_set.meta_data
-        ]
+        return result_set.description
 
     def description(self, query_id):
         return self._executor.submit(self._description, query_id)
@@ -56,13 +60,16 @@ class AsyncCursor(BaseCursor):
     def _collect_result_set(self, query_id):
         query_execution = self._poll(query_id)
         return AthenaResultSet(
-            self._connection, self._converter, query_execution, self.arraysize,
+            self._connection, self._converter, query_execution, self._arraysize,
             self.retry_exceptions, self.retry_attempt, self.retry_multiplier,
             self.retry_max_delay, self.retry_exponential_base)
 
     def execute(self, operation, parameters=None):
         query_id = self._execute(operation, parameters)
         return query_id, self._executor.submit(self._collect_result_set, query_id)
+
+    def executemany(self, operation, seq_of_parameters):
+        raise NotSupportedError
 
     def cancel(self, query_id):
         return self._executor.submit(self._cancel, query_id)
